+++ lib/src/models/embed_constant.dart
@@ -1,3 +1,4 @@
 const Duration kDefaultEmbedHtmlCacheLifeSpan = Duration(days: 7);
 const Duration kDefaultEmbedLoadTimeout = Duration(seconds: 20);
 const double kDefaultMaxScrollableEmbedHeight = 400.0;
+const double kDefaultHeightUpdateDeltaThreshold = 2.0;